requirements.txt
# requirements.txt
beautifulsoup4
requests
sentence-transformers
faiss-cpu
fastapi
uvicorn[standard]
numpy
python-dotenv
google-generativeai
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0